[VX-16] Created new files V2X_car_state_check to encapsulate the car state checking so as not to clog up the power or can files.
